+++ R/AcqOptimizer.R
@@ -243,7 +243,9 @@ AcqOptimizer = R6Class("AcqOptimizer",
     #' @field param_set ([paradox::ParamSet])\cr
     #'   Set of hyperparameters.
     param_set = function(rhs) {
-      assert_ro_binding(rhs)
+      if (!missing(rhs) && !identical(rhs, private$.param_set)) {
+        stop("$param_set is read-only.")
+      }
       private$.param_set
     }
   ),