+++ R/OptimizerMbo.R
@@ -187,7 +187,7 @@ OptimizerMbo = R6Class("OptimizerMbo",
       }
 
       if (is.null(self$acq_optimizer)) {
-        self$acq_optimizer = default_acqoptimizer(self$acq_function)
+        self$acq_optimizer = default_acqoptimizer(self$acq_function, inst)
       }
 
       if (is.null(self$result_assigner)) {