+++ tests/testthat/test_AcqFunctionStochasticCB.R
@@ -4,6 +4,8 @@ test_that("AcqFunctionStochasticCB works in defaults", {
   skip_if_not(redis_available())
   flush_redis()
 
+  options(bbotk.debug = TRUE)
+
   mirai::daemons(1L)
   rush::rush_plan(n_workers = 1L, worker_type = "remote")
   instance = oi_async(
@@ -14,10 +16,13 @@ test_that("AcqFunctionStochasticCB works in defaults", {
 
   acq_function = acqf("stochastic_cb")
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
@@ -42,10 +47,13 @@ test_that("AcqFunctionStochasticCB works with uniform sampling", {
 
   acq_function = acqf("stochastic_cb", distribution = "uniform", min_lambda = 1, max_lambda = 3)
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5L,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
@@ -66,15 +74,18 @@ test_that("AcqFunctionStochasticCB works with exponential sampling", {
   instance = oi_async(
     objective = OBJ_2D,
     search_space = PS_2D,
-    terminator = trm("evals", n_evals = 10L),
+    terminator = trm("evals", n_evals = 50L),
   )
 
   acq_function = acqf("stochastic_cb", distribution = "exponential", lambda = 1.96)
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
@@ -101,10 +112,13 @@ test_that("AcqFunctionStochasticCB works with lambda decay", {
 
   acq_function = acqf("stochastic_cb", rate = 0.5)
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5L,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
@@ -131,10 +145,13 @@ test_that("AcqFunctionStochasticCB works with periodic lambda decay", {
 
   acq_function = acqf("stochastic_cb", rate = 0.5, period = 2)
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5L,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)