+++ tests/testthat/test_AcqFunctionStochasticEI.R
@@ -14,14 +14,17 @@ test_that("AcqFunctionStochasticEI works in defaults", {
 
   acq_function = acqf("stochastic_ei")
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5L,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
-  expect_names(names(instance$archive$data), must.include = c(".already_evaluated", "acq_ei", "acq_epsilon_0", "acq_epsilon"))
+  expect_names(names(instance$archive$data), must.include = c("acq_ei", ".already_evaluated", "acq_epsilon_0", "acq_epsilon"))
   expect_numeric(-instance$archive$data$acq_epsilon, sorted = TRUE)
 
   expect_rush_reset(instance$rush)
@@ -43,14 +46,17 @@ test_that("AcqFunctionStochasticEI works with multiple workers", {
 
   acq_function = acqf("stochastic_ei")
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5L,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
-  expect_names(names(instance$archive$data), must.include = c(".already_evaluated", "acq_ei", "acq_epsilon_0", "acq_epsilon"))
+  expect_names(names(instance$archive$data), must.include = c("acq_ei", ".already_evaluated", "acq_epsilon_0", "acq_epsilon"))
 
   expect_rush_reset(instance$rush)
 })
@@ -72,14 +78,17 @@ test_that("AcqFunctionStochasticEI works with periodic epsilon decay", {
 
   acq_function = acqf("stochastic_ei", rate = 0.5, period = 2)
 
+  acq_optimizer = acqo(opt("random_search"), trm("evals", n_evals = 10L), acq_function = acq_function)
+
   optimizer = opt("async_mbo",
     design_function = "sobol",
     design_size = 5L,
-    acq_function = acq_function)
+    acq_function = acq_function,
+    acq_optimizer = acq_optimizer)
 
   expect_data_table(optimizer$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
-  expect_names(names(instance$archive$data), must.include = c(".already_evaluated", "acq_ei", "acq_epsilon_0", "acq_epsilon"))
+  expect_names(names(instance$archive$data), must.include = c("acq_ei", ".already_evaluated", "acq_epsilon_0", "acq_epsilon"))
   expect_numeric(unique(instance$archive$data$acq_epsilon), len = 3L)
 
   expect_rush_reset(instance$rush)