+++ tests/testthat/test_AcqOptimizerRandomSearch.R
@@ -6,15 +6,11 @@ test_that("AcqOptimizerRandomSearch works", {
   surrogate = srlrn(REGR_KM_DETERM, archive = instance$archive)
   acqfun = acqf("ei", surrogate = surrogate)
   acqopt = AcqOptimizerRandomSearch$new(acq_function = acqfun)
-  acqopt$param_set$set_values(max_fevals = 200L)
+  acqopt$param_set$set_values(n_evals = 200L)
   acqfun$surrogate$update()
   acqfun$update()
 
   expect_data_table(acqopt$optimize(), nrows = 1L)
-  expect_list(acqopt$state)
-  expect_names(names(acqopt$state), must.include = "iteration_1")
-  expect_class(acqopt$state$iteration_1$model, "nloptr")
-  expect_true(acqopt$state$iteration_1$model$iterations <= 200L)
 })
 
 test_that("AcqOptimizerRandomSearch works with 2D", {
@@ -30,10 +26,6 @@ test_that("AcqOptimizerRandomSearch works with 2D", {
   acqfun$update()
 
   expect_data_table(acqopt$optimize(), nrows = 1L)
-  expect_list(acqopt$state)
-  expect_names(names(acqopt$state), must.include = "iteration_1")
-  expect_class(acqopt$state$iteration_1$model, "nloptr")
-  expect_true(acqopt$state$iteration_1$model$iterations <= 200L)
 })
 
 test_that("AcqOptimizerRandomSearch works with instance", {
@@ -47,6 +39,6 @@ test_that("AcqOptimizerRandomSearch works with instance", {
   acqopt$param_set$set_values(n_evals = 10L)
 
   optimizer = opt("mbo", acq_optimizer = acqopt, acq_function = acqfun, surrogate = surrogate)
-  expect_data_table(optimizer$optimize(instance), nrow = 1L)
+  expect_data_table(optimizer$optimize(instance), nrows = 1L)
 })