+++ tests/testthat/test_TunerAsyncMbo.R
@@ -23,7 +23,7 @@ test_that("TunerAsyncMbo works", {
 
   expect_data_table(tuner$optimize(instance), nrows = 1L)
   expect_data_table(instance$archive$data, min.rows = 10L)
-  expect_names(names(instance$archive$data), must.include = c("acq_cb", ".already_evaluated", "acq_lambda_0", "acq_lambda"))
+  expect_names(names(instance$archive$data), must.include = "acq_cb")
 
   expect_rush_reset(instance$rush)
 })