+++ tests/testthat/test_mbo_defaults.R
@@ -21,84 +21,125 @@ test_that("default_surrogate", {
   surrogate = default_surrogate(MAKE_INST_1D())
   expect_r6(surrogate, "SurrogateLearner")
   expect_r6(surrogate$learner, "LearnerRegrKM")
-  expect_equal_sorted(surrogate$learner$param_set$values,
-    list(covtype = "matern5_2", optim.method = "gen", control = list(trace = FALSE), nugget.stability = 1e-08))
+  expect_equal_sorted(surrogate$learner$param_set$values, list(
+    scaling = FALSE,
+    covtype = "matern5_2",
+    optim.method = "gen",
+    control = list(trace = FALSE),
+    nugget.stability = 1e-08))
   expect_equal(surrogate$learner$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner$fallback, "LearnerRegrRanger")
+  expect_r6(surrogate$output_trafo, "OutputTrafoLog")
 
   # singlecrit all numeric, noisy
   surrogate = default_surrogate(MAKE_INST_1D_NOISY())
   expect_r6(surrogate, "SurrogateLearner")
   expect_r6(surrogate$learner, "LearnerRegrKM")
-  expect_equal_sorted(surrogate$learner$param_set$values,
-    list(covtype = "matern5_2", optim.method = "gen", control = list(trace = FALSE), nugget.estim = TRUE, jitter = 1e-12))
+  expect_equal_sorted(surrogate$learner$param_set$values, list(
+    scaling = FALSE,
+    covtype = "matern5_2",
+    optim.method = "gen",
+    control = list(trace = FALSE),
+    nugget.estim = TRUE,
+    jitter = 1e-12))
   expect_equal(surrogate$learner$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner$fallback, "LearnerRegrRanger")
+  expect_r6(surrogate$output_trafo, "OutputTrafoLog")
 
   # twocrit all numeric, deterministic
   surrogate = default_surrogate(MAKE_INST(OBJ_1D_2, search_space = PS_1D))
   expect_r6(surrogate, "SurrogateLearnerCollection")
   expect_list(surrogate$learner, types = "LearnerRegrKM")
-  expect_equal_sorted(surrogate$learner[[1L]]$param_set$values,
-    list(covtype = "matern5_2", optim.method = "gen", control = list(trace = FALSE), nugget.stability = 1e-08))
+  expect_equal_sorted(surrogate$learner[[1L]]$param_set$values, list(
+    scaling = FALSE,
+    covtype = "matern5_2",
+    optim.method = "gen",
+    control = list(trace = FALSE),
+    nugget.stability = 1e-08))
   expect_equal(surrogate$learner[[1L]]$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner[[1L]]$fallback, "LearnerRegrRanger")
   expect_equal(surrogate$learner[[1L]]$param_set$values, surrogate$learner[[2L]]$param_set$values)
   expect_equal(surrogate$learner[[1L]]$encapsulation, surrogate$learner[[2L]]$encapsulation)
-  suppressWarnings(expect_equal(surrogate$learner[[1L]]$fallback, surrogate$learner[[2L]]$fallback))
+  expect_equal(surrogate$learner[[1L]]$fallback, surrogate$learner[[2L]]$fallback)
+  expect_r6(surrogate$output_trafo, "OutputTrafoLog")
 
   # twocrit all numeric, noisy
   surrogate = default_surrogate(MAKE_INST(OBJ_1D_2_NOISY, search_space = PS_1D))
   expect_r6(surrogate, "SurrogateLearnerCollection")
   expect_list(surrogate$learner, types = "LearnerRegrKM")
-  expect_equal_sorted(surrogate$learner[[1L]]$param_set$values,
-    list(covtype = "matern5_2", optim.method = "gen", control = list(trace = FALSE), nugget.estim = TRUE, jitter = 1e-12))
+  expect_equal_sorted(surrogate$learner[[1L]]$param_set$values, list(
+    scaling = FALSE,
+    covtype = "matern5_2",
+    optim.method = "gen",
+    control = list(trace = FALSE),
+    nugget.estim = TRUE,
+    jitter = 1e-12))
   expect_equal(surrogate$learner[[1L]]$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner[[1L]]$fallback, "LearnerRegrRanger")
   expect_equal(surrogate$learner[[1L]]$param_set$values, surrogate$learner[[2L]]$param_set$values)
   expect_equal(surrogate$learner[[1L]]$encapsulation, surrogate$learner[[2L]]$encapsulation)
   expect_equal(surrogate$learner[[1L]]$fallback, surrogate$learner[[2L]]$fallback)
+  expect_r6(surrogate$output_trafo, "OutputTrafoLog")
 
   # singlecrit mixed input
   surrogate = default_surrogate(MAKE_INST(OBJ_1D_MIXED, search_space = PS_1D_MIXED))
   expect_r6(surrogate, "SurrogateLearner")
   expect_r6(surrogate$learner, "LearnerRegrRanger")
-  expect_equal_sorted(surrogate$learner$param_set$values[c("num.threads", "num.trees", "keep.inbag", "se.method")],
-    list(num.threads = 1L, num.trees = 100L, keep.inbag = TRUE, se.method = "jack"))
+  expect_equal_sorted(surrogate$learner$param_set$values, list(
+    num.threads = 1L,
+    num.trees = 500L,
+    se.method = "jack",
+    splitrule = "variance",
+    keep.inbag = TRUE,
+    sample.fraction = 1,
+    min.node.size = 3,
+    min.bucket = 3,
+    mtry.ratio = 5 / 6,
+    sigma2.threshold = 1e-2))
   expect_equal(surrogate$learner$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner$fallback, "LearnerRegrRanger")
+  expect_null(surrogate$output_trafo)
 
   # twocrit mixed input
   surrogate = default_surrogate(MAKE_INST(OBJ_1D_2_MIXED, search_space = PS_1D_MIXED))
   expect_r6(surrogate, "SurrogateLearnerCollection")
   expect_list(surrogate$learner, types = "LearnerRegrRanger")
-  expect_equal_sorted(surrogate$learner[[1L]]$param_set$values[c("num.threads", "num.trees", "keep.inbag", "se.method")],
-    list(num.threads = 1L, num.trees = 100L, keep.inbag = TRUE, se.method = "jack"))
+  expect_equal_sorted(surrogate$learner[[1]]$param_set$values, list(
+    num.threads = 1L,
+    num.trees = 500L,
+    se.method = "jack",
+    splitrule = "variance",
+    keep.inbag = TRUE,
+    sample.fraction = 1,
+    min.node.size = 3,
+    min.bucket = 3,
+    mtry.ratio = 5 / 6,
+    sigma2.threshold = 1e-2))
   expect_equal(surrogate$learner[[1L]]$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner[[1L]]$fallback, "LearnerRegrRanger")
   expect_equal(surrogate$learner[[1L]]$param_set$values, surrogate$learner[[2L]]$param_set$values)
   expect_equal(surrogate$learner[[1L]]$encapsulation, surrogate$learner[[2L]]$encapsulation)
   expect_equal(surrogate$learner[[1L]]$fallback, surrogate$learner[[2L]]$fallback)
+  expect_null(surrogate$output_trafo)
 
   # singlecrit mixed input deps
   surrogate = default_surrogate(MAKE_INST(OBJ_1D_MIXED, search_space = PS_1D_MIXED_DEPS))
   expect_r6(surrogate, "SurrogateLearner")
-  expect_r6(surrogate$learner, "GraphLearner")
-  expect_equal(surrogate$learner$graph$ids(), c("imputesample", "imputeoor", "colapply", "regr.ranger"))
-  list_expected = list(imputesample.affect_columns = mlr3pipelines::selector_type("logical"),
-     imputeoor.min = TRUE,
-     imputeoor.offset = 1,
-     imputeoor.multiplier = 3,
-     imputeoor.affect_columns = mlr3pipelines::selector_type(c("integer", "numeric", "character", "factor", "ordered")),
-     colapply.applicator = as.factor,
-     colapply.affect_columns = mlr3pipelines::selector_type("character"),
-     regr.ranger.num.threads = 1L,
-     regr.ranger.num.trees = 100L,
-     regr.ranger.keep.inbag = TRUE,
-     regr.ranger.se.method = "jack")
-  expect_equal_sorted(surrogate$learner$param_set$values[names(list_expected)], list_expected)
-
-  expect_r6(surrogate$learner$fallback, "LearnerRegrFeatureless")
+  expect_r6(surrogate$learner, "LearnerRegrRanger")
+  expect_equal_sorted(surrogate$learner$param_set$values, list(
+    num.threads = 1L,
+    num.trees = 500L,
+    se.method = "jack",
+    splitrule = "variance",
+    keep.inbag = TRUE,
+    sample.fraction = 1,
+    min.node.size = 3,
+    min.bucket = 3,
+    mtry.ratio = 5 / 6,
+    sigma2.threshold = 1e-2))
+  expect_equal(surrogate$learner$encapsulation, c(train = "evaluate", predict = "evaluate"))
+  expect_r6(surrogate$learner$fallback, "LearnerRegrRanger")
+  expect_null(surrogate$output_trafo)
 
   # specify own learner, specify n_objectives, twocrit all numeric, deterministic
   surrogate = default_surrogate(MAKE_INST(OBJ_1D_2, search_space = PS_1D), learner = lrn("regr.featureless"), n_learner = 1L)
@@ -112,14 +153,22 @@ test_that("default_surrogate", {
 
 test_that("default_acqfunction", {
   instance = MAKE_INST_1D()
-  acq = default_acqfunction(instance)
-  expect_r6(acq, "AcqFunctionEI")
+  acq_function = default_acqfunction(instance)
+  expect_r6(acq_function, "AcqFunctionEILog")
+
+  instance = MAKE_INST(OBJ_1D_MIXED, search_space = PS_1D_MIXED)
+  acq_function = default_acqfunction(instance)
+  expect_r6(acq_function, "AcqFunctionCB")
+
+  instance = MAKE_INST(OBJ_1D_2, search_space = PS_1D)
+  acq_function = default_acqfunction(instance)
+  expect_r6(acq_function, "AcqFunctionSmsEgo")
 })
 
 test_that("default_acqoptimizer", {
-  acqopt = default_acqoptimizer(default_acqfunction(MAKE_INST_1D()))
-  expect_r6(acqopt, "AcqOptimizer")
-  expect_r6(acqopt$optimizer, "OptimizerBatchRandomSearch")
+  acq_function = default_acqfunction(MAKE_INST_1D())
+  acqopt = default_acqoptimizer(acq_function, MAKE_INST_1D())
+  expect_r6(acqopt, "AcqOptimizerLocalSearch")
 })
 
 test_that("default_result_assigner", {
@@ -155,12 +204,13 @@ test_that("stability and defaults", {
   learner = LearnerRegrError$new()
   learner$encapsulate("evaluate", lrn("regr.ranger", num.trees = 10L, keep.inbag = TRUE, se.method = "jack"))
   surrogate = default_surrogate(instance, learner = learner, n_learner = 1L)
+  surrogate$output_trafo = OutputTrafoLog$new(invert_posterior = FALSE)
   expect_r6(surrogate, "SurrogateLearner")
   expect_r6(surrogate$learner, "LearnerRegrError")
   expect_equal(surrogate$learner$encapsulation, c(train = "evaluate", predict = "evaluate"))
   expect_r6(surrogate$learner$fallback, "LearnerRegrRanger")
   acq_function = default_acqfunction(instance)
-  expect_r6(acq_function, "AcqFunctionEI")
+  expect_r6(acq_function, "AcqFunctionEILog")
   acq_optimizer = acqo(opt("random_search", batch_size = 2L), terminator = trm("evals", n_evals = 2L))
   acq_optimizer$param_set$values$logging_level = "info"
   expect_r6(acq_optimizer, "AcqOptimizer")