+++ NEWS.md
@@ -1,7 +1,8 @@
 # mlr3mbo (development version)
 
 * feat: Added `mlr_acqoptimizers` dictionary.
-* feat: Added `AcqOptimizerCmaes`, `AcqOptimizerDirect`, `AcqOptimizerLbfgsb`, `AcqOptimizerLocalSearch`, and `AcqOptimizerRandomSearch`.
+* perf: Added `AcqOptimizerDirect`, `AcqOptimizerLbfgsb`, `AcqOptimizerLocalSearch`, and `AcqOptimizerRandomSearch`.
+* feat: Added
 
 # mlr3mbo 0.3.3